requirements/dev.txt
-r base.txt

faker==6.6.3
pytest~=5.4.1
pytest-aiohttp~=0.3.0
pytest-cov==2.8.1
prospector==1.3.0